+++ neatrl/src/neatrl/__init__.py
@@ -1,6 +1,6 @@
 from .dqn import train_dqn
 from .dueling_dqn import train_dueling_dqn
 from .reinforce import train_reinforce
-from .rnd import train_rnd
+from .rnd import train_ppo_rnd, train_ppo_rnd_cnn
 
-__all__ = ["train_dqn", "train_reinforce", "train_dueling_dqn", "train_rnd"]
+__all__ = ["train_dqn", "train_reinforce", "train_dueling_dqn", "train_ppo_rnd", "train_ppo_rnd_cnn"]